chore: put extensions API to the separate repository
