--- crates/extension_api/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "extension_api"
-version = "0.1.0"
-edition = "2024"
-rust-version.workspace = true
-
-[dependencies]
-anyhow = { workspace = true }
-php_sys = { workspace = true }
-tokio = { workspace = true, features = ["sync"] }
-
-[lints]
-workspace = true
