+++ crates/integration_tests/Cargo.toml
@@ -15,6 +15,6 @@ log = "0.4"
 
 [dev-dependencies]
 anyhow = "1"
-extension_api = { path = "../extension_api" }
+extension_api = { workspace = true }
 extension_host = { path = "../extension_host" }
 tokio = { version = "1", features = ["macros", "rt", "rt-multi-thread"] }
